terminal: New term.js with fix focus and selection fixes
term.js doesn't blink the cursor anymore, because that caused the
selection to vanish. Blink the cursor with a css animation instead.

In addition, show the cursor as an empty outline when the terminal
doesn't have focus.

Also remove the one remaining call to showCursor(), which we removed in
term.js-cockpit 0.0.10 (it's a private fork). There was never a need for
calling that function, as we set cursorHidden to false and refreshed the
terminal anyways.

Fixes #7143
Closes #7983