storage: Fix get_parent for cleartext devices
The CryptoBackingDevice property is on the Block interface, not on
Encrypted.

This makes a job spinner appear for a drive when formatting a
encrypted filesystem on it and makes links to encrypted devices work
as intended.

But we keep the old behavior of not showing a spinner on the backing
device.  There is already a spinner on the cleartext device, and
showing two spinners next to each other is confusing.

This is done via a more general rule: Only the immediate object of a
job gets a spinner, and its top-most parent object.

Closes #7955