test: Fix timeout handling in Browser.wait()
This method is usually called on external commands, not on JavaScript
code inside the browser. This causes the arm_timeout() handler set in
PhantomJS to not trigger until the "global" and fixed one-minute timeout
in phantom-driver.js's `step()` kicks in. This would also not work at
all if the predicate function itself hangs.

Thus use a timeout on the test side, using `alarm()`.

Closes #8015
Reviewed-by: Peter <petervo@redhat.com>