lib: Use new 'post_hook' to set unit string on plots
The setup_hook is too early for this. We need the grid to be setup as
well to know the actual ranges of the axes.

Previously, we would show the unit string for the previous refresh of
a plot, which is almost always correct.  But when scrolling and
zooming around in archive data, it is easy to get the wrong unit
string.

Closes #8060
Reviewed-by: Peter <petervo@redhat.com>