test: Fix another race in check-terminal
Handling key presses and rendering their effect in the terminal is not
synchronous with `Browser.key_press()`. It could happen that the line
showed the keys half-typed (`admin@localhost ~]$ clea`) which would
already satisfy the following wait condition. So first wait until the
key presses got processed, then wait for the clear command to take
effect.

Closes #8182