ws: Create self-signed certificates with subjectAltName
Chromium insists on certificates having subjectAltName.  With this the
certificate can be imported into your ~/.pki to make Chromium accept it:

    openssl s_client -connect 127.0.0.1:9090 < /dev/null | \
        sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' > /tmp/cert.pem
    certutil -d sql:$HOME/.pki/nssdb -A -t "TC,C,P" -n cockpit -i /tmp/cert.pem

Add localhost and 127.0.0.1 as most common cases for now. Other cases
can be added in the future upon request (e. g. for official private IP
ranges).

This does not change `create_temp_file()`, just moves it further up in
the file so that we can use it from `openssl_make_dummy_cert()`.