packagekit: Fix crash on security updates without CVE URLs
If an update is marked as type "security" but has no CVE URLs,
`insertCommas()` is called on an empty list. This does not work with the
`.reduce()` function and it crashes with

    TypeError: reduce of empty array with no initial value

Fix this by checking for an empty or one-element list in
`insertCommas()`. (The latter because `Array.reduce` returns the element
itself for an one-element list)

Reproduce this in a test case with `severity="security"` and no CVE
URLs. This only works on yum, as PackageKit's apt backend does not
support severities, and the dnf backend yields wrong values
(https://bugs.freedesktop.org/show_bug.cgi?id=101070). For that, move
the backend detection from `TestAutoUpdates` to the parent `PackageCase`
class.

Fixes #8375
Closes #8377