bots: Implement cluster based machine learning
This attempts to cluster log items related to similarity and then
classify whether new test logs fit into those clusters. The clustering
is unsupervised, and currently uses DBSCAN to accomplish this.
The classification currently uses nearest neighbor techniques.

We use distances to tell us whether two items are similar or not.
These distances are currently calculated via normalized compression
distance in ncd.py

We run this code in addition to the older Neural Network machine
learning so we can compare the difference in effectiveness.

The machine learning uploads the clusters it finds to the logs
when it performs its job.

The tests-policy script has been updated to print out the likelyhood
that a test failure is a flake.

Closes #8380